+++ app/src/androidTest/java/ch/epfl/sdp/peakar/user/FriendsActivityTest.java
@@ -27,7 +27,7 @@ import ch.epfl.sdp.peakar.user.friends.FriendItemActivity;
 import ch.epfl.sdp.peakar.user.friends.FriendsActivity;
 import ch.epfl.sdp.peakar.user.services.AuthAccount;
 import ch.epfl.sdp.peakar.user.services.AuthService;
-import ch.epfl.sdp.peakar.user.services.providers.firebase.FirebaseAuthService;
+import ch.epfl.sdp.peakar.user.services.FirebaseAuthService;
 import ch.epfl.sdp.peakar.user.friends.RemoteFriendItem;
 
 import static androidx.test.espresso.Espresso.onData;
