+++ app/src/main/java/ch/epfl/sdp/peakar/collection/CollectionActivity.java
@@ -14,7 +14,7 @@ import ch.epfl.sdp.peakar.R;
 import ch.epfl.sdp.peakar.user.score.ScoringConstants;
 import ch.epfl.sdp.peakar.user.services.AuthAccount;
 import ch.epfl.sdp.peakar.user.services.AuthService;
-import ch.epfl.sdp.peakar.user.services.providers.firebase.FirebaseAuthService;
+import ch.epfl.sdp.peakar.user.services.FirebaseAuthService;
 import ch.epfl.sdp.peakar.utils.ToolbarHandler;
 
 public class CollectionActivity extends AppCompatActivity {
